\name{EnsDb.Drerio.v104}
\docType{package}

\alias{EnsDb.Drerio.v104-package}
\alias{EnsDb.Drerio.v104}
\alias{EnsDb.Drerio.v104}


\title{Ensembl based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Thu Nov  3 11:56:31 2022 and based on the 104
}

\author{O Davalos}



\examples{
## load the library
##library(EnsDb.Drerio.v104)
## list the contents that are loaded into memory
ls('package:EnsDb.Drerio.v104')
## show the db object that is loaded by calling it's name
EnsDb.Drerio.v104

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
